phpredis: rename tests for php redis suite following folder based versioning
